fix: composeWithDevTools should not be used in production
