+++ root.reducer.js
@@ -1,4 +1,9 @@
 import { combineReducers } from 'redux';
+import { persistReducer } from 'redux-persist';
+import createEncryptor from 'redux-persist-transform-encrypt';
+import md5 from 'md5';
+import DeviceInfo from 'react-native-device-info';
+import AsyncStorage from '@react-native-community/async-storage';
 import { authReducer } from 'auth';
 import { userReducer } from 'user';
 import { repositoryReducer } from 'repository';
@@ -7,7 +12,17 @@ import { issueReducer } from 'issue';
 import { notificationsReducer } from 'notifications';
 import { entities, pagination } from 'api/reducers';
 
-export const rootReducer = combineReducers({
+const encryptor = createEncryptor({
+  secretKey: md5(DeviceInfo.getUniqueID()),
+});
+const rootPersistConfig = {
+  key: 'root',
+  storage: AsyncStorage,
+  transforms: [encryptor],
+  whitelist: ['auth'],
+};
+
+const rootReducer = combineReducers({
   auth: authReducer,
   user: userReducer,
   repository: repositoryReducer,
@@ -17,3 +32,5 @@ export const rootReducer = combineReducers({
   entities,
   pagination,
 });
+
+export default persistReducer(rootPersistConfig, rootReducer);
